tea/pyproject.toml
[project]
name = "tea"
version = "0.1.0"
description = "Personal teagarden manager"
readme = "README.md"
authors = [
    { name = "Bertold Sedlak", email = "berti.sedlak@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "pyyaml>=6.0.3",
    "typer>=0.21.1",
]

[project.scripts]
tea = "tea.cli:app"

[build-system]
requires = ["uv_build>=0.9.18,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "black>=25.12.0",
    "isort>=7.0.0",
    "pytest>=9.0.2",
    "ruff>=0.14.11",
]
